usb: k1x_udc_core: use semaphore to prime new dtds
According to datasheet:

ATDTW: Add dTD TripWire – RW. Default = 0b.

This bit is used as a semaphore to ensure the to proper addition
of a new dTD to an active (primed) endpoint’s linked list. This
bit is set and cleared by software. This bit shall also be
cleared by hardware when its state machine is in a hazard region
for which adding a dTD to a primed endpoint may go unrecognized.

This logic is already implemented in mv_udc_core.c, fsl_udc_core.c
and chipidea/udc.c

Change-Id: Ide5b5f3ff0ca04216fa0e6889eb18eb9cd3f7819
Signed-off-by: Junzhong Pan <junzhong.pan@spacemit.com>