k1: usb: f_ncm: set bMaxBurst to 0 for spacemit k1
It's observed large bMaxBurst on spacemit k1 cause some XHCI host
having issue when doing iperf3 test.

An Error scene of epin on bus trace looks like this:
  Host TP ACK SeqN=22 NumP=8 Retry=N
  Dev  DP     SeqN=23 Data=1024B **(Not Acked DP)**
  Host TP ACK SeqN=23 NumP=7 (Asked for DP SeqN=23)
  Dev  DP	    SeqN=24 Data=600B **(Aborted DPP)**
  Host TP ACK SeqN=23 NumP=7 Retry=Y (Asked for DP SeqN=23 Again)
  Dev  TP NRDY
  Dev  TP ERDY        NumP=16
  ITPs ...
And epout looks like this:
  Dev ERDY, Host DP, Dev TP ACK, Host DP, Dev TP ACK, ITPs...

It seems the host fails to send ACK again on the epin and fails to
send more data on the epout. Setting bMaxBurst=0 could mitigate the
issue, host error is not excluded. More investigation will be
performed, for better compatibility, disable bursting for now.

The networking throughput is not decreasing compared to bMaxBurst=15.

Change-Id: I16f5adba4871874a56932a705772ec35a088c344
Signed-off-by: Junzhong Pan <junzhong.pan@spacemit.com>