k1: support user-space load and store misaligned on io memory area
The PCIe BAR area must be mapped as IO attribute for GPU VRAM area,
but, there may be some misaligned access in user-space program. Then,
it will trigger 5# exception when misaligned load and 7# excepiton
when misaligned store. And user-space program will crash because of
the 5# exception and 7# exception can't be fixed.
So, we try to process the user-space io area misligned access firstly
when the 5# exception and 7# exception are caught

Change-Id: Iac78bcdb8505dd113dcc4deed782e78016ddbf9b